requirements.txt
keboola.component==1.3.7
keboola.utils
snowflake-connector-python
git+https://github.com/keboola/sapi-python-client.git
mock
freezegun
dataconf